EMP-31O Moved formatting to nunjucks filters
