+++ server/utils/crmDisplayConfigValidation.ts
@@ -21,7 +21,7 @@ const schema = Joi.object({
               {
                 label: Joi.string().optional().allow(''),
                 apiField: Joi.string().required(),
-                format: Joi.string().valid('currency', 'date', 'time').optional(),
+                type: Joi.string().valid('currency', 'date', 'time').optional(),
               },
               {
                 subHeading: Joi.string().required(),
