+++ server/utils/nunjucksSetup.ts
@@ -5,6 +5,7 @@ import express from 'express'
 import { initialiseName } from './utils'
 import { ApplicationInfo } from '../applicationInfo'
 import config from '../config'
+import { formatCurrency, formatDate, formatTime } from './crmFieldFormatter'
 
 const production = process.env.NODE_ENV === 'production'
 
@@ -41,4 +42,7 @@ export default function nunjucksSetup(app: express.Express, applicationInfo: App
   )
 
   njkEnv.addFilter('initialiseName', initialiseName)
+  njkEnv.addFilter('formatCurrency', formatCurrency)
+  njkEnv.addFilter('formatDate', formatDate)
+  njkEnv.addFilter('formatTime', formatTime)
 }
